refactor: move to hooks
