+++ src/admin/components/elements/BlockEditor/hooks/useBlockEditor.ts
@@ -1,0 +1,62 @@
+import { Extension } from '@tiptap/core';
+import CharacterCount from '@tiptap/extension-character-count';
+import Heading from '@tiptap/extension-heading';
+import TiptapLink from '@tiptap/extension-link';
+import Placeholder from '@tiptap/extension-placeholder';
+import TaskItem from '@tiptap/extension-task-item';
+import Underline from '@tiptap/extension-underline';
+import { Content, useEditor } from '@tiptap/react';
+import { StarterKit } from '@tiptap/starter-kit';
+import { useTranslation } from 'react-i18next';
+export { Heading } from '@tiptap/extension-heading';
+export { TaskItem } from '@tiptap/extension-task-item';
+
+export const useBlockEditor = ({
+  initialContent,
+  ref,
+}: {
+  initialContent: Content;
+  ref: React.RefObject<HTMLButtonElement>;
+}) => {
+  const { t } = useTranslation();
+
+  const editor = useEditor({
+    extensions: [
+      StarterKit.configure({
+        heading: {
+          levels: [2, 3],
+          HTMLAttributes: {
+            class: 'heading',
+          },
+        },
+      }),
+      Underline,
+      CharacterCount,
+      Placeholder.configure({ placeholder: t('write_the_text') }),
+      TiptapLink.configure({
+        openOnClick: false,
+        autolink: true,
+        defaultProtocol: 'https',
+      }),
+      Heading.configure({
+        levels: [1, 2, 3, 4, 5, 6],
+      }),
+      TaskItem.configure({
+        nested: true,
+      }),
+      Extension.create({
+        addKeyboardShortcuts() {
+          return {
+            'Mod-s': () => {
+              ref.current?.click();
+              return true;
+            },
+          };
+        },
+      }),
+    ],
+    content: initialContent,
+  });
+
+  return { editor };
+};
